Move memory map to sys::boot module
