+++ src/sys/mem/mod.rs
@@ -9,6 +9,7 @@ pub use paging::{
 };
 pub use phys::{phys_addr, PhysBuf};
 
+use crate::sys::boot::MemoryMap;
 use crate::sys::pic;
 
 use core::sync::atomic::{AtomicUsize, Ordering};
@@ -21,61 +22,6 @@ static mut MAPPER: Once<OffsetPageTable<'static>> = Once::new();
 
 static PHYS_MEM_OFFSET: Once<u64> = Once::new();
 static MEMORY_SIZE: AtomicUsize = AtomicUsize::new(0);
-
-const MAX_REGIONS: usize = 32;
-
-#[repr(u32)]
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum MemoryRegionType {
-    Usable,
-    Reserved,
-    AcpiUsable,
-    AcpiReserved,
-    Defective,
-    Custom(u32),
-}
-
-#[derive(Copy, Clone, Debug)]
-pub struct MemoryRegion {
-    addr: u64,
-    size: u64,
-    kind: MemoryRegionType,
-}
-
-impl MemoryRegion {
-    pub fn new(addr: u64, size: u64, kind: MemoryRegionType) -> Self {
-        Self { addr, size, kind }
-    }
-}
-
-#[derive(Copy, Clone, Debug)]
-pub struct MemoryMap {
-    regions: [MemoryRegion; MAX_REGIONS],
-    len: usize,
-}
-
-impl MemoryMap {
-    pub fn new() -> Self {
-        let empty = MemoryRegion::new(0, 0, MemoryRegionType::Reserved);
-        Self {
-            regions: [empty; MAX_REGIONS],
-            len: 0,
-        }
-    }
-
-    pub fn add(&mut self, region: MemoryRegion) {
-        self.regions[self.len] = region;
-        self.len += 1;
-    }
-
-    pub fn as_slice(&self) -> &[MemoryRegion] {
-        &self.regions[..self.len]
-    }
-
-    pub fn iter(&self) -> core::slice::Iter<'_, MemoryRegion> {
-        self.as_slice().iter()
-    }
-}
 
 pub fn init(memory_map: &MemoryMap, offset: u64) {
     // Keep the timer interrupt to have accurate boot time measurement but mask
